fix editTask
